+++ NEWS.md
@@ -20,3 +20,4 @@
 * Fix error causing 3d arrows to not display
 * Update the way point stress and residual is calculated for the stress and residual error table
 * Added option to plot.acmap to indicate points outside of plot limits with an arrow
+* Fixed error when performing dimensional annealing